Rewrite shadow piercing descendant combinators (#574)
I'm not very happy with the loss of compartmentalization here,
where previously all the skin-related css could live in a single file.
I'm not sure this is scalable, but nobody else has added any other
skins, so this seems fine for now.

An alternative here is that we could add css variables for all
the properties in a timer-bar that are modified by lippe.css right
now, but that's not very extensible either.