raidboss: fix timeline loading race
There's a race between all the conditions for ReloadTimelines
and cactbotReadDataFiles.  If cactbotReadDataFiles is slow,
then timelines are missing and errors get printed.

raidboss can't wait for the data files entirely, because
it needs to handle a bunch of other events being sent.

Handle this by not loading timelines until data files are ready
and explicitly reloading once they are.